Add files via upload
In this section, there is a report on how to install and set up the zabbix system and network monitoring tool in the Linux operating system. This report is in Farsi language, but the video of setting up, installing and using it is prepared in English and is placed in the link below.
In this tutorial, how to install zabbix server and zabbix agent is taught.
You can install zabbix server on the target system where you want to measure various metrics. Also install zabbix agent on any other system or the same system. and view the analysis of different metrics collected from the target system by Zabbix server.